+++ rbx_reflection/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Added `ReflectionDatabase::superclasses_iter`
+* Added `ReflectionDatabase::has_superclass`, like `Instance:IsA(ClassName)`
 
 ## 4.7.0 (2024-08-22)
 * Update to rbx_types 1.10